OP_CHECKMERKLEBRANCH: New opcode to validate root, leaf, and path for a Merkle tree, plus one with -VERIFY semantics.
Useful for implementing tree signatures and MAST (in conjunction with some sort of OP_EVAL). The opcode checks that the root generated from the given path and leaf hash value matches the reference value provided, using the fast Merkle branching code. Has a built-in depth limitation of 32-levels for the path.